fix address name after village/structure settle
